Intial NavBar Styling Complete
