+++ References/OldNavbar/Navbar.stories.js
@@ -1,0 +1,16 @@
+import React from "react"
+import Navbar from "./Navbar"
+
+export default {
+  title: "Navbar",
+}
+
+export const standard = () => (
+  <div
+    style={{
+      background: "black",
+    }}
+  >
+    <Navbar />
+  </div>
+)
