+++ References/Props Notes.mdx
@@ -1,0 +1,50 @@
+# Props Reference
+
+## Styling with Functional Props in React (w/o styled-components or emotion)
+
+### Defining props in style objects
+
+```jsx
+const TestTheme = {
+  colors: {
+    me: "#c6fcb3",
+  },
+}
+```
+
+### Setting props in the child component
+
+```jsx
+<Foo color={props.color} />
+```
+
+### Calling props in the parent component
+
+```jsx
+<Foo color={TestTheme.colors.me} />
+```
+
+## Styling via emotion
+
+```javascript
+const Button = styled.button`
+  font-size: 1em;
+  color: ${props => props.theme.main};
+  border: 2px solid ${props => props.theme.main};
+`
+
+const theme = {
+  main: "mediumseagreen",
+}
+render(
+  <div>
+    <Button theme={{ main: "royalblue" }}>Ad hoc theme</Button>
+    <ThemeProvider theme={theme}>
+      <div>
+        <Button>Themed</Button>
+        <Button theme={{ main: "darkorange" }}>Overidden</Button>
+      </div>
+    </ThemeProvider>
+  </div>
+)
+```
