+++ src/components/molecules/Header/Header.jsx
@@ -3,7 +3,7 @@ import { jsx } from "theme-ui"
 import React from "react"
 import { Link } from "gatsby"
 import PropTypes from "prop-types"
-import NavMenu from "../../navigation/NavMenu"
+import NavBar from "../../navigation/NavBar"
 import Logo from "../../atoms/Logo"
 import { StaticQuery, graphql } from "gatsby"
 
@@ -22,15 +22,15 @@ const Header = ({ props }) => (
       <header
         {...props}
         sx={{
-          bg: "black",
           display: "flex",
           alignItems: "center",
-          justifyContent: "space-between",
+          justifyContent: theme => `${theme.header.justifyContent}`,
+          backgroundColor: theme => `${theme.header.backgroundColor}`,
         }}
       >
         {/* <h1>{data.site.siteMetadata.title}</h1> */}
         <Logo />
-        <NavMenu />
+        <NavBar />
       </header>
     )}
   />
