--- src/components/navigation/NavMenu/index.js
@@ -1,3 +1,0 @@
-import NavMenu from './NavMenu';
-
-export default NavMenu;
