+++ src/components/navigation/Navbar/Navbar.jsx
@@ -1,50 +1,57 @@
 /** @jsx jsx */
-import React from "react"
-import { Box, Flex, jsx, NavLink } from "theme-ui"
-import { Link } from "@reach/router"
 
-const Navbar = ({ image, title, description }) => (
-  <Flex as="nav" px={2} alignItems="center">
-    <Box mx="auto" />
-    <NavLink>
-      <Link
-        to="/"
-        sx={{
-          variant: "links.mainNav",
-        }}
-      >
-        services
-      </Link>
-    </NavLink>
-    <NavLink>
-      <Link
-        to="/"
-        sx={{
-          variant: "links.mainNav",
-        }}
-      >
-        insights
-      </Link>
-    </NavLink>
-    <NavLink>
-      <Link
-        to="/"
-        sx={{
-          variant: "links.mainNav",
-        }}
-      >
-        profile
-      </Link>
-    </NavLink>
-  </Flex>
+import { jsx } from "theme-ui"
+import React from "react"
+import { StaticQuery, graphql, Link } from "gatsby"
+import PropTypes from "prop-types"
+import { Box, Flex } from "theme-ui"
+
+const NavBar = props => (
+  <StaticQuery
+    query={graphql`
+      query SiteQuery {
+        site {
+          siteMetadata {
+            title
+            menuLinks {
+              name
+              link
+            }
+          }
+        }
+      }
+    `}
+    render={data => (
+      <React.Fragment>
+        <Box
+          as="nav"
+          sx={{
+            variant: "navBar.vertical.box",
+          }}
+        >
+          <Flex
+            as="ul"
+            sx={{
+              variant: "navBar.vertical.ul",
+            }}
+          >
+            {data.site.siteMetadata.menuLinks.map(link => (
+              <li
+                key={link.name}
+                sx={{
+                  variant: "navBar.vertical.li",
+                }}
+              >
+                <Link to={link.link} sx={{ variant: "navBar.vertical.link" }}>
+                  {link.name}
+                </Link>
+              </li>
+            ))}
+          </Flex>
+        </Box>
+      </React.Fragment>
+    )}
+  />
 )
 
-Navbar.propTypes = {
-  // bla: PropTypes.string,
-}
-
-Navbar.defaultProps = {
-  // bla: 'test',
-}
-
-export default Navbar
+export default NavBar
