--- src/components/navigation/Navbar/Navbar.styles.js
@@ -1,6 +1,0 @@
-// import styled from 'styled-components';
-
-// export const Test = styled.div`
-//  display: flex;
-// `;
-//
