+++ src/components/organisms/Layout/Layout.jsx
@@ -4,6 +4,10 @@ import React from "react"
 import Header from "../../molecules/Header"
 import Footer from "../../molecules/Footer"
 
+const PageHeader = {
+  background: "black",
+}
+
 const Layout = ({ children }) => (
   <Flex
     sx={{
