+++ src/gatsby-plugin-theme-ui/index.js
@@ -1,9 +1,9 @@
 import React from "react"
 import typography from "../utils/typography"
 import { merge } from "theme-ui"
-import navMenu from "./navMenu"
+import navBar from "./navBar"
 
-export default merge(navMenu, {
+export default merge(navBar, {
   breakpoints: ["40em", "52em", "64em"],
   space: [0, 4, 8, 16, 32, 64, 128, 256, 512],
   fonts: {
