+++ src/pages/index.js
@@ -1,4 +1,3 @@
 import React from "react"
 import { IndexPage } from "../components/pages"
-
 export default () => <IndexPage />
